fix(dedup): address CodeRabbit review findings on PR #110
Three findings from CodeRabbit review, all APPLY:

1. dtype consistency (embedder.py line 106): empty-corpus return path used
   float32 while the non-empty path returns float64 (sklearn default via
   toarray()). Changed np.float32 → np.float64 to keep both paths consistent
   and avoid downstream concatenation/caching dtype mismatches.

2. Missing assertions in edge-case tests (test_dedup_semantic.py):
   test_single_document_corpus and test_zero_document_corpus only checked 4 of
   6 keys; added median_similarity and std_similarity assertions — exactly the
   fields that would have raised/NaN'd under the old implementation.

3. RUF002 unicode in docstrings: replaced '×' (U+00D7) with plain 'x' in two
   test docstrings to satisfy Ruff's ambiguous-unicode-character rule.

+++ src/services/deduplication/embedder.py
@@ -103,7 +103,7 @@ class DocumentEmbedder:
         """
         if not documents:
             logger.warning("Empty document list provided")
-            return np.empty((0, self.max_features), dtype=np.float32)
+            return np.empty((0, self.max_features), dtype=np.float64)
 
         logger.info(f"Fitting vectorizer on {len(documents)} documents")
